o  fixed build controller
git-svn-id: https://svn.apache.org/repos/asf/maven/continuum/branches/id-refactor@497930 13f79535-47bb-0310-9956-ffa450edef68

+++ continuum-core/src/main/java/org/apache/maven/continuum/buildcontroller/BuildController.java
@@ -29,6 +29,6 @@ public interface BuildController
 {
     String ROLE = BuildController.class.getName();
 
-    void build( int projectId, int buildDefinitionId, int trigger )
+    void build( long projectId, long buildDefinitionId, int trigger )
         throws TaskExecutionException;
 }
